pyproject.toml
[tool.poetry]
name = "dajeobot"
version = "0.1.0"
description = ""
authors = ["Danil <headcrabj1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-telegram-bot = "^20.3"
peewee = "^3.16.2"
schedule = "^1.2.0"
psycopg2 = "^2.9.6"
python-dateutil = "^2.8.2"
babel = "^2.12.1"
pyinstaller = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
